feat: dcm2bids optional zip processing
The ability to process scans within zip folders should not be default behaviour because users may have already unzipped the scans. With this approach the zips are not processed.

Future: if no scans are found no in zips, then maybe there should be a fallback to look for zips instead. This will prevent double up processing while still providing utility unprompted by the user.