fix: dcm2bids more robust name and cohort parsing
The StudyDescription can vary quite dramatically across subjects in studies. This causes problems with the parsing of subject and session name and cohort. In cases where the formatting of StudyDescription is too short and likely truncated, the PatientName is used. This may need to be adjusted if future acquisitions use different terminology.